+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 ------------------
 
 - Cleanup docs
+- Cleanup metadata.xml file
+- Change SwissSign Idp endpoint URL
+- Show SAML2 metadata file configuration option
+- Patch pySAML2 so that POST is default binding in metadata configuration
 
 0.1b1 (2010-10-22)
 ------------------